requirements.txt
asyncio
# chromadb
deepgram-sdk
deepgram-captions
# flask
# flask-restx
flake8
flake8-bugbear
jq
langchain
langchain-openai
openai
openapi-schema-pydantic
python-dateutil
python-dotenv
regex
requests
httpx
# youtube-transcript-api
pika
pymongo
pytest